pyproject.toml
[project]
name = "simcore"
version = "0.2.0"
description = "PISA simulation core - drives the simulation loop via gRPC"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pisa-api>=0.1.20",
    "pyyaml>=6.0.3",
]

[tool.uv.sources]
pisa-api = { git = "https://github.com/PISA-Hut/pisa-api.git" }
